完成了用户信息的下载和上传，声音的数值和slider的对应关系还需要修改，新手引导还需要修改

+++ LetterCard/Assets/LetterCardFiles/Scripts/Card/FunctionCard.cs
@@ -299,7 +299,7 @@ public class FunctionCard : Card
                 else
                 {
                     int tmpCoinCount = Random.Range(1, GameManager.Instance.CanPlayZeroCardScore);
-                    GameEntrance.instance.CoinCount += (uint)tmpCoinCount;
+                    GameEntrance.instance.CoinCount += tmpCoinCount;
                     ShowTipManager.instance.ShowTip("¿Õ°×Êé¾í", tmpCoinCount);
                 }
                 break;
